feat: add Begin/End markers and fix pathfinding interaction
feat: support Begin/End markers in pathfinding tests

- dynamic start/end via map markers
- fix movement blocking on End tile
- align pathfinding rules with non-blocking structures

+++ Assets/Scripts/Game/Map/Stucture/Utility/UtilityMap.cs
@@ -908,13 +908,13 @@ public class UtilityMap : StructureMap<UtilityMap>
                 {
                     Structure structure = GetStructure(new Vector3Int(x, y));
                     if(structure != null)
-                        list.Add(new MapCellData(x, y, 0, (int)structure.Type));
+                        list.Add(new MapCellData(x, y, 0, structure.Type));
                 }
                 else
                 {
                     Structure structure = GetStructure(new Vector3Int(x, y));
                     if (structure != null)
-                        list.Add(new MapCellData(x, y, 0, (int)structure.Type));
+                        list.Add(new MapCellData(x, y, 0, structure.Type));
                 }
             }
         }
